+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tasks moved to separate tasks tab
 - Upgraded dependencies (major versions)
 - Use new SearchBar widget
+- Use new SearchBar widget on Habits page
 
 ## [0.8.349] - 2023-05-16
 ### Changed: